change test back to serial execution
